Add stream in places  favorite btn
