+++ places/lib/ui/screen/sight_list_screen/widgets/add_sight_btn.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:places/res/res.dart';
